% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_swap.R
\name{check_swap}
\alias{check_swap}
\title{Check the result of the swapping}
\usage{
check_swap(risks, donors, res_swap)
}
\arguments{
\item{risks}{data.table}

\item{donors}{data.table}

\item{res_swap}{data.table result of the swap}
}
\value{
list
}
\description{
Check the result of the swapping
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
donors <- create_data_example(n)
risks <- donors[sample(1:n, ceiling(n*0.07)),][, scope_risk := "geo"]
donors_drawn <- draw_donors_multi_geo_multi_sim(donors, risks, l_similar = list(c("edu", "sex", "age"), c("age", "sex"), c("sex")), geo_levels = "geo")
check_res <- check_swap(risks, donors, donors_drawn)
}
