% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_examples.R
\name{create_data_example}
\alias{create_data_example}
\title{Create a dataset of fake individuals with some characteristics}
\usage{
create_data_example(n = 1e+05, add_geo = FALSE)
}
\arguments{
\item{n}{number of individuals}
}
\value{
data.table object of n rows and 7 variables
(one ident, one geo var, three cat vars and two var for the risk)
}
\description{
Create a dataset of fake individuals with some characteristics
}
\examples{
library(data.table)
set.seed(123)
individuals <- create_data_example(n = 1e5)
individuals[, .N, by = .(edu, sex)]
individuals[, .N, by = .(sex, age)]
individuals[, .N, by = .(geo)]
individuals <- create_data_example(n = 1e5, add_geo = TRUE)
individuals[, .N, by = .(geo2)]
individuals[, .N, by = .(scope_risk)]
}
