% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors}
\alias{draw_donors}
\title{Title}
\usage{
draw_donors(donors, risks, stats_risks, geo_level)
}
\arguments{
\item{donors}{data.table}

\item{risks}{data.table}

\item{stats_risks}{data.table}

\item{geo_level}{character vector}
}
\value{
a data.table
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
donors <- create_data_example(n)
risks <- donors[sample(1:n, ceiling(n*0.07)),]
stats_risks <- summary_risk(donors, risks, similar = c("edu", "sex"), geo_level = "geo")
donors_drawn <- draw_donors(donors, risks, stats_risks, "geo")
}
