% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_multi_geo_multi_sim}
\alias{draw_donors_multi_geo_multi_sim}
\title{Title}
\usage{
draw_donors_multi_geo_multi_sim(
  donors,
  risks,
  l_similar,
  geo_levels,
  parallel = FALSE,
  n_cores = 2
)
}
\arguments{
\item{donors}{data.table}

\item{risks}{data.table}

\item{l_similar}{list of character vector}

\item{geo_levels}{character vector}
}
\value{
a data.table
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
data <- create_data_example(n)
data_prep <- prepare_data(data, "is_risky", "scope_risk", "ident", c("edu", "sex", "age"), "geo")
risks <- data_prep$risks
donors <- data_prep$donors
donors_drawn <- draw_donors_multi_geo_multi_sim(donors, risks, l_similar = list(c("edu", "sex", "age"), c("age", "sex"), c("sex")), geo_levels = "geo")
check_res <- check_swap(risks, donors, donors_drawn)
}
