% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_multi_sim}
\alias{draw_donors_multi_sim}
\title{Title}
\usage{
draw_donors_multi_sim(donors, risks, l_similar, geo_level)
}
\arguments{
\item{l_similar}{list of character vectors}

\item{geo_level}{}
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
donors <- create_data_example(n)
risks <- donors[sample(1:n, ceiling(n*0.07)),]
donors_drawn <- draw_donors_multi_sim(donors, risks, l_similar = list(c("edu", "sex", "age"), c("age", "sex"), c("sex")), geo_level = "geo")
}
