% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_one_geo_multi_sim}
\alias{draw_donors_one_geo_multi_sim}
\title{Title}
\usage{
draw_donors_one_geo_multi_sim(
  donors_geo,
  risks_geo,
  l_similar,
  geo_level,
  geo_level_sup = NULL
)
}
\arguments{
\item{donors_geo}{data.table}

\item{risks_geo}{data.table}

\item{l_similar}{list of character vectors}

\item{geo_level}{character vctor}

\item{geo_level_sup}{character vector}
}
\value{
a data.table
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
data <- create_data_example(n)
data_prep <- prepare_data(data, "is_risky", "scope_risk", "ident", c("edu", "sex", "age"), "geo")
risks <- data_prep$risks
donors <- data_prep$donors
donors_drawn <- draw_donors_one_geo_multi_sim(donors, risks, l_similar = list(c("edu", "sex", "age"), c("age", "sex"), c("sex")), geo_level = "geo")
}
