% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_one_geo_one_sim2}
\alias{draw_donors_one_geo_one_sim2}
\title{Title}
\usage{
draw_donors_one_geo_one_sim2(
  donors_geo,
  risks_geo,
  stats_risks,
  similar,
  geo_level,
  geo_level_sup = NULL
)
}
\arguments{
\item{geo_level_sup}{}
}
\value{
data.table
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
data <- create_data_example(n, add_geo = TRUE)
data_prep <- prepare_data(data, "is_risky", "scope_risk", "ident", c("edu", "sex", "age"), c("geo","geo2"))
risks <- data_prep$risks
donors <- data_prep$donors
stats_risks <- summary_risk(donors, risks, similar = c("edu", "sex"), geo_level = "geo", geo_level_sup = "geo2")
donors_drawn <- draw_donors_one_geo_one_sim2(donors, risks, stats_risks, similar = c("edu", "sex"), geo_level = "geo", geo_level_sup = "geo2")
}
