% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Title}
\usage{
prepare_data(data, var_risk, var_scope_risk, var_ident, vars_similar, vars_geo)
}
\arguments{
\item{data}{data.table}

\item{var_risk}{name of a boolean var}

\item{var_scope_risk}{name of a character/num var}

\item{var_ident}{name of a char or num var}

\item{vars_similar}{names of all the similar vars (char or numeric variables)}

\item{vars_geo}{names of all the geo vars (char or num vars)}
}
\value{
list of two data.tables: donors and risks
}
\description{
Title
}
\examples{
set.seed(123)
n = 1e3
data <- create_data_example(n)
data_prep <- prepare_data(data, "is_risky", "scope_risk", "ident", c("edu", "sex", "age"), "geo")
}
