% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_one_geo_multi_sim}
\alias{draw_donors_one_geo_multi_sim}
\title{Title}
\usage{
draw_donors_one_geo_multi_sim(
  donors_geo,
  risks_geo,
  l_similar,
  geo_level,
  geo_level_sup = NULL
)
}
\arguments{
\item{donors_geo}{data.table}

\item{risks_geo}{data.table}

\item{l_similar}{list of character vectors}

\item{geo_level}{character vctor}

\item{geo_level_sup}{character vector}
}
\value{
a data.table
}
\description{
Title
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
donors <- create_data_example(n)
risks <- donors[sample(1:n, ceiling(n*0.07)),]
donors_geo = data.table::copy(donors); risks_geo = data.table::copy(risks)
donors_drawn <- draw_donors_one_geo_multi_sim(donors_geo, risks_geo, l_similar = list(c("edu", "sex", "age"), c("age", "sex"), c("sex")), geo_level = "geo")
}
