% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_donors.R
\name{draw_donors_one_geo_one_sim}
\alias{draw_donors_one_geo_one_sim}
\title{Match similar risks-donors at one geographic level}
\usage{
draw_donors_one_geo_one_sim(
  donors_geo,
  risks_geo,
  stats_risks,
  similar,
  geo_level,
  geo_level_sup = NULL
)
}
\arguments{
\item{donors_geo}{data.table}

\item{risks_geo}{data.table}

\item{stats_risks}{data.table}

\item{similar}{character vector}

\item{geo_level}{character vector}

\item{geo_level_sup}{character vector}
}
\value{
a data.table
}
\description{
Match similar risks-donors at one geographic level
}
\examples{
library(data.table)
set.seed(123)
n = 1e3
donors <- create_data_example(n)
risks <- donors[sample(1:n, ceiling(n*0.07)),]
stats_risks <- summary_risk(donors, risks, similar = c("edu", "sex"), geo_level = "geo")
donors_geo = data.table::copy(donors); risks_geo = data.table::copy(risks)
donors_drawn <- draw_donors_one_geo_one_sim(donors, risks, stats_risks, similar = c("edu", "sex"), geo_level = "geo")
}
